pyproject.toml
[tool.poetry]
name = "pigsqueeze"
version = "1.0.0"
description = "A library to write and read arbitrary data to and from image files. You probably already know why you need it."
authors = ["Tim Wedde <timwedde@icloud.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/timwedde/pigsqueeze"
homepage = "https://github.com/timwedde/pigsqueeze"

[tool.poetry.scripts]
psz = "pigsqueeze.cli:cli"

[tool.poetry.dependencies]
python = "^3.8"
plum-py = "^0.4.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest = "^6.2.1"
black = "^20.8b1"
pre-commit = "^2.9.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
